Introduce AbstractApiHandler
